phase-10d T4: assembly across the grid; identity cell reproduces 10c BIT-EXACTLY
Verified on real data, not only in the T2 replay: the (K=0, d=0, min_prints=2,
sep=hard_break) cell reproduces all 170,722 of 10c's sub-bursts with identical start_ns,
end_ns, n_prints and duration_s, and identical move_share. 2,629,020 intervals and 130 ok
(event,kernel) cells both match 10c exactly.

T4c break-cause census, first measurement of this split in the programme: 1,299 of 170,592
run breaks (0.761%) involve an ok=False interval. Per kernel 1.01% / 0.57% / 0.18% at
2 / 8 / 32 min. At interval level the ok=False share is 11.9%, higher because those
intervals cluster inside long separator runs that break once.

Upstream config-hash discrepancy, recorded not worked around: Stage 1's digest, REPORT.md
and s1_t0_denominator.json all carry config_hash 998c2461, which is config/phase_10c.json
as of commit 0f079a9. Commit 39ec87e edited that config INSIDE Stage 1, before 692d9d0
produced the T1 artifacts, so the recorded hash is stale by one commit. True value
17124203 (LF) / a0943e2a (raw CRLF). Also noted: 10c's cfg_hash() hashes raw file bytes
and is line-ending sensitive -- Stage 0 recorded an LF-form hash, Stage 0b and Stage 1
CRLF-form. 10d asserts both forms. 10c artifacts are not edited.

+++ research/phase_10d/assemble.py
@@ -150,7 +150,10 @@ def assemble(norm: np.ndarray, ok: np.ndarray, ts: np.ndarray, thr: float,
         "n_merges": np.array([m["merges"] for m in merged], dtype=np.int64)[keep],
         "start_ns": ts[start[keep]],
         "end_ns": ts[end[keep] + 1],
-        "duration_s": (ts[end[keep] + 1].astype(np.float64) - ts[start[keep]].astype(np.float64)) / 1e9,
+        # int64 subtraction FIRST, then scale -- subtracting two float64-cast nanosecond
+        # epochs loses ~256 ns of precision at 1e18 magnitudes and would stop the identity
+        # cell reproducing 10c's durations exactly.
+        "duration_s": (ts[end[keep] + 1] - ts[start[keep]]).astype(np.float64) / 1e9,
         "n_objects": int(keep.sum()),
         "n_objects_before_floor": int(len(merged)),
         "n_deleted_by_floor": int((~keep).sum()),
